fix: move withRouter in separate directory
